refactor: split test cases page and remove any
